Versão 2.1.1
Atualização código da tabela de vendas.

+++ src/controller/Produto.java
@@ -37,7 +37,7 @@ public class Produto {
         return codigo;
     }
 
-    public void setCodigo(int codigo) {
+    public void setCodigo(long codigo) {
         this.codigo = codigo;
     }
